make legacy export more robust

+++ bin/export.py
@@ -235,7 +235,8 @@ def build_problem_zip(problem: Problem, output: Path) -> bool:
     if config.args.legacy and languages:
         for type in PdfType:
             file = export_dir / type.path(languages[0], ".pdf").name
-            file.rename(remove_language_pdf_suffix(file, languages[0]))
+            if file.exists():
+                file.rename(remove_language_pdf_suffix(file, languages[0]))
     else:
         for language in languages:
             for type in PdfType:
@@ -333,19 +334,20 @@ def build_problem_zip(problem: Problem, output: Path) -> bool:
         if (export_dir / "statement").exists():
             (export_dir / "statement").rename(export_dir / "problem_statement")
         for d in ["solution", "problem_slide"]:
-            for f in list(util.glob(problem.path, f"{d}/*")):
-                if f.is_file():
-                    out = Path("problem_statement") / f.relative_to(problem.path / d)
-                    if out.exists():
-                        message(
-                            f"Can not export {f.relative_to(problem.path)} as {out}",
-                            "Zip",
-                            output,
-                            color_type=MessageType.WARN,
-                        )
-                    else:
-                        add_file(out, f)
-            shutil.rmtree(export_dir / d)
+            if (export_dir / d).is_dir():
+                for f in list(util.glob(problem.path, f"{d}/*")):
+                    if f.is_file():
+                        out = Path("problem_statement") / f.relative_to(problem.path / d)
+                        if out.exists():
+                            message(
+                                f"Can not export {f.relative_to(problem.path)} as {out}",
+                                "Zip",
+                                output,
+                                color_type=MessageType.WARN,
+                            )
+                        else:
+                            add_file(out, f)
+                shutil.rmtree(export_dir / d)
 
     # handle yaml updates
     yaml_path.unlink()
